+++ pipeline/resolve.go
@@ -1,19 +1,24 @@
 package pipeline
 
 import (
+	"context"
+	"errors"
 	"fmt"
 
 	"github.com/streamingfast/bstream"
+	"github.com/streamingfast/bstream/hub"
+	"github.com/streamingfast/dstore"
 	pbssinternal "github.com/streamingfast/substreams/pb/sf/substreams/intern/v2"
 	pbsubstreamsrpc "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2"
 	"github.com/streamingfast/substreams/reqctx"
+	"go.uber.org/zap"
 	grpccodes "google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
 type getRecentFinalBlockFunc func() (uint64, error)
 
-func BuildRequestDetails(request *pbsubstreamsrpc.Request, getRecentFinalBlock getRecentFinalBlockFunc) (req *reqctx.RequestDetails, err error) {
+func BuildRequestDetails(ctx context.Context, request *pbsubstreamsrpc.Request, getRecentFinalBlock getRecentFinalBlockFunc, resolveCursor CursorResolver) (req *reqctx.RequestDetails, undoSignal *pbsubstreamsrpc.BlockUndoSignal, err error) {
 	req = &reqctx.RequestDetails{
 		Modules:                             request.Modules,
 		OutputModule:                        request.OutputModule,
@@ -21,19 +26,20 @@ func BuildRequestDetails(request *pbsubstreamsrpc.Request, getRecentFinalBlock g
 		ProductionMode:                      request.ProductionMode,
 		StopBlockNum:                        request.StopBlockNum,
 	}
-	req.RequestStartBlockNum, err = resolveStartBlockNum(request)
+
+	req.RequestStartBlockNum, undoSignal, err = resolveStartBlockNum(ctx, request, resolveCursor)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
 	linearHandoff, err := computeLiveHandoffBlockNum(request.ProductionMode, req.RequestStartBlockNum, request.StopBlockNum, getRecentFinalBlock)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
 	req.LinearHandoffBlockNum = linearHandoff
 
-	return req, nil
+	return
 }
 
 func BuildRequestDetailsFromSubrequest(request *pbssinternal.ProcessRangeRequest) (req *reqctx.RequestDetails) {
@@ -70,7 +76,8 @@ func computeLiveHandoffBlockNum(productionMode bool, startBlock, stopBlock uint6
 	return minOf(startBlock, maxHandoff), nil
 }
 
-func resolveStartBlockNum(req *pbsubstreamsrpc.Request) (uint64, error) {
+// resolveStartBlockNum will occasionally modify or remove the cursor inside the request
+func resolveStartBlockNum(ctx context.Context, req *pbsubstreamsrpc.Request, resolveCursor CursorResolver) (uint64, *pbsubstreamsrpc.BlockUndoSignal, error) {
 	// TODO(abourget): a caller will need to verify that, if there's a cursor.Step that is New or Undo,
 	// then we need to validate that we are returning not only a number, but an ID,
 	// We then need to sync from a known finalized Snapshot's block, down to the potentially
@@ -79,27 +86,91 @@ func resolveStartBlockNum(req *pbsubstreamsrpc.Request) (uint64, error) {
 	// and everything is irreversible.
 
 	if req.StartBlockNum < 0 {
-		return 0, status.Error(grpccodes.InvalidArgument, "start block num must be positive")
+		return 0, nil, status.Error(grpccodes.InvalidArgument, "start block num must be positive")
 	}
 
 	if req.StartCursor == "" {
-		return uint64(req.StartBlockNum), nil
+		return uint64(req.StartBlockNum), nil, nil
 	}
 
 	cursor, err := bstream.CursorFromOpaque(req.StartCursor)
 	if err != nil {
-		return 0, status.Errorf(grpccodes.InvalidArgument, "invalid start cursor %q: %s", cursor, err.Error())
+		return 0, nil, status.Errorf(grpccodes.InvalidArgument, "invalid start cursor %q: %s", cursor, err.Error())
 	}
-	if cursor.Step.Matches(bstream.StepIrreversible) {
-		return cursor.Block.Num() + 1, nil // this block was the last sent to the customer
+
+	if cursor.IsOnFinalBlock() {
+		nextBlock := cursor.Block.Num() + 1
+		req.StartCursor = ""
+		return nextBlock, nil, nil
 	}
-	if cursor.Step.Matches(bstream.StepNew) {
-		return cursor.Block.Num() + 1, nil // this block was the last sent to the customer
+
+	lastValidBlock, head, err := resolveCursor(ctx, cursor)
+	if err != nil {
+		return 0, nil, status.Errorf(grpccodes.InvalidArgument, "cannot resolve startCursor %q: %s", cursor, err.Error())
 	}
-	if cursor.Step.Matches(bstream.StepUndo) {
-		return cursor.Block.Num(), nil
+	var undoSignal *pbsubstreamsrpc.BlockUndoSignal
+	if lastValidBlock.Num() != cursor.Block.Num() {
+		validCursor := &bstream.Cursor{
+			Step:      bstream.StepNew,
+			Block:     lastValidBlock,
+			LIB:       cursor.LIB,
+			HeadBlock: head,
+		}
+		req.StartCursor = validCursor.ToOpaque()
+
+		undoSignal = &pbsubstreamsrpc.BlockUndoSignal{
+			LastValidClock:  blockRefToClock(lastValidBlock),
+			LastValidCursor: req.StartCursor,
+		}
 	}
-	return 0, fmt.Errorf("invalid start cursor step")
+
+	return lastValidBlock.Num() + 1, undoSignal, nil
+}
+
+type CursorResolver func(context.Context, *bstream.Cursor) (lastValidBlock, currentHead bstream.BlockRef, err error)
+
+type junctionBlockGetter struct {
+	foundBlock  bstream.BlockRef
+	currentHead bstream.BlockRef
+}
+
+var ErrJunctionFound = errors.New("junction block found")
+
+func (j *junctionBlockGetter) ProcessBlock(block *bstream.Block, obj interface{}) error {
+	j.foundBlock = obj.(bstream.Stepable).ReorgJunctionBlock()
+	j.currentHead = obj.(bstream.Cursorable).Cursor().HeadBlock
+	return ErrJunctionFound
+}
+
+func NewCursorResolver(hub *hub.ForkableHub, mergedBlocksStore, forkedBlocksStore dstore.Store) CursorResolver {
+
+	return func(ctx context.Context, cursor *bstream.Cursor) (lastValidBlock, currentHead bstream.BlockRef, err error) {
+		jctBlkGetter := &junctionBlockGetter{}
+		src := hub.SourceFromCursor(cursor, jctBlkGetter)
+		if src == nil { // block is out of reversible segment
+			src = bstream.NewFileSourceFromCursor(mergedBlocksStore, forkedBlocksStore, cursor, jctBlkGetter, zap.NewNop())
+		}
+		if src == nil {
+			return nil, nil, fmt.Errorf("cannot get source to resolve cursor")
+		}
+
+		src.Run()
+		select {
+		case <-ctx.Done():
+			return nil, nil, ctx.Err()
+		case <-src.Terminated():
+		}
+
+		if !errors.Is(src.Err(), ErrJunctionFound) {
+			headBlock := cursor.HeadBlock
+			if headNum, headID, _, _, err := hub.HeadInfo(); err == nil {
+				headBlock = bstream.NewBlockRef(headID, headNum)
+			}
+			return cursor.LIB, headBlock, nil
+		}
+
+		return jctBlkGetter.foundBlock, jctBlkGetter.currentHead, nil
+	}
 }
 
 func minOf(a, b uint64) uint64 {