+++ pipeline/resolve_test.go
@@ -1,6 +1,7 @@
 package pipeline
 
 import (
+	"context"
 	"fmt"
 	"testing"
 
@@ -14,10 +15,12 @@ import (
 
 func Test_resolveStartBlockNum(t *testing.T) {
 	tests := []struct {
-		name             string
-		req              *pbsubstreamsrpc.Request
-		expectedBlockNum uint64
-		wantErr          bool
+		name               string
+		req                *pbsubstreamsrpc.Request
+		expectedBlockNum   uint64
+		wantErr            bool
+		cursorResolverArgs []interface{}
+		wantUndoLastBlock  bstream.BlockRef
 	}{
 		{
 			name: "invalid cursor step",
@@ -34,7 +37,7 @@ func Test_resolveStartBlockNum(t *testing.T) {
 			wantErr:          true,
 		},
 		{
-			name: "step undo",
+			name: "step undo", // substreams customers should not receive these cursors now, kept for backwards compatibility
 			req: &pbsubstreamsrpc.Request{
 				StartBlockNum: 10,
 				StartCursor: (&bstream.Cursor{
@@ -46,6 +49,10 @@ func Test_resolveStartBlockNum(t *testing.T) {
 			},
 			expectedBlockNum: 10,
 			wantErr:          false,
+			cursorResolverArgs: []interface{}{
+				"c1:2:10:10a:9:9a", bstream.NewBlockRef("9a", 9), bstream.NewBlockRef("9a", 9), nil,
+			},
+			wantUndoLastBlock: bstream.NewBlockRef("9a", 9),
 		},
 		{
 			name: "step new",
@@ -62,13 +69,31 @@ func Test_resolveStartBlockNum(t *testing.T) {
 			wantErr:          false,
 		},
 		{
-			name: "step irreversible",
+			name: "step new on forked cursor",
+			req: &pbsubstreamsrpc.Request{
+				StartBlockNum: 10,
+				StartCursor: (&bstream.Cursor{
+					Step:      bstream.StepNew,
+					Block:     bstream.NewBlockRef("10a", 10),
+					LIB:       bstream.NewBlockRef("6a", 6),
+					HeadBlock: bstream.NewBlockRef("10a", 10),
+				}).ToOpaque(),
+			},
+			expectedBlockNum: 9,
+			wantErr:          false,
+			cursorResolverArgs: []interface{}{
+				"c1:1:10:10a:6:6a", bstream.NewBlockRef("8a", 8), bstream.NewBlockRef("11a", 11), nil,
+			},
+			wantUndoLastBlock: bstream.NewBlockRef("8a", 8),
+		},
+		{
+			name: "step irreversible", // substreams should not receive these cursors now, kept for backwards compatibility
 			req: &pbsubstreamsrpc.Request{
 				StartBlockNum: 10,
 				StartCursor: (&bstream.Cursor{
 					Step:      bstream.StepIrreversible,
 					Block:     bstream.NewBlockRef("10a", 10),
-					LIB:       bstream.NewBlockRef("9a", 9),
+					LIB:       bstream.NewBlockRef("10a", 10),
 					HeadBlock: bstream.NewBlockRef("10a", 10),
 				}).ToOpaque(),
 			},
@@ -92,7 +117,15 @@ func Test_resolveStartBlockNum(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := resolveStartBlockNum(tt.req)
+			got, undoSignal, err := resolveStartBlockNum(context.Background(), tt.req, newTestCursorResolver(tt.cursorResolverArgs...).resolveCursor)
+			if tt.wantUndoLastBlock != nil {
+				require.NotNil(t, undoSignal)
+				assert.Equal(t, tt.wantUndoLastBlock.ID(), undoSignal.LastValidClock.Id)
+				assert.Equal(t, tt.wantUndoLastBlock.Num(), undoSignal.LastValidClock.Number)
+			} else {
+				assert.Nil(t, undoSignal)
+			}
+
 			if (err != nil) != tt.wantErr {
 				t.Errorf("resolveStartBlockNum() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -153,7 +186,8 @@ func Test_computeLiveHandoffBlockNum(t *testing.T) {
 }
 
 func TestBuildRequestDetails(t *testing.T) {
-	req, err := BuildRequestDetails(
+	req, _, err := BuildRequestDetails(
+		context.Background(),
 		&pbsubstreamsrpc.Request{
 			StartBlockNum:  10,
 			ProductionMode: false,
@@ -162,12 +196,15 @@ func TestBuildRequestDetails(t *testing.T) {
 		func() (uint64, error) {
 			assert.True(t, true, "should pass here")
 			return 999, nil
-		})
+		},
+		newTestCursorResolver().resolveCursor,
+	)
 	require.NoError(t, err)
 	assert.Equal(t, 10, int(req.RequestStartBlockNum))
 	assert.Equal(t, 10, int(req.LinearHandoffBlockNum))
 
-	req, err = BuildRequestDetails(
+	req, _, err = BuildRequestDetails(
+		context.Background(),
 		&pbsubstreamsrpc.Request{
 			StartBlockNum:  10,
 			ProductionMode: true,
@@ -175,7 +212,9 @@ func TestBuildRequestDetails(t *testing.T) {
 		},
 		func() (uint64, error) {
 			return 999, nil
-		})
+		},
+		newTestCursorResolver().resolveCursor,
+	)
 	require.NoError(t, err)
 	assert.Equal(t, 10, int(req.RequestStartBlockNum))
 	assert.Equal(t, 999, int(req.LinearHandoffBlockNum))